[Refactor #1]: Add index to properly export components
