+++ src/App.js
@@ -1,20 +1,17 @@
 import Mockman from "mockman-js";
 import { Route, Routes } from "react-router-dom";
 import "./App.css";
-import Cart from "./Pages/Cart/Cart";
-import Home from "./Pages/Home/Home";
-import ProductList from "./Pages/ProductList/ProductList";
-import Wishlist from "./Pages/Wishlist/Wishlist";
+import { Home, Cart, Wishlist, ProductList } from "./Pages";
 
 function App() {
   return (
     <div className="App">
       <Routes>
-        <Route path="/" element={<Home/>}/>
-        <Route path="/cart" element={<Cart/>}/>
-        <Route path="/wishlist" element={<Wishlist/>}/>
-        <Route path="/products" element={<ProductList/>}/>
-        <Route path="/mockman" element={<Mockman/>}/>
+        <Route path="/" element={<Home />} />
+        <Route path="/cart" element={<Cart />} />
+        <Route path="/wishlist" element={<Wishlist />} />
+        <Route path="/products" element={<ProductList />} />
+        <Route path="/mockman" element={<Mockman />} />
       </Routes>
     </div>
   );
