+++ src/Components/index.js
@@ -1,0 +1,7 @@
+import CategoryCard from "./CategoryCard/CategoryCard";
+import CartCard from "./CartCard/CartCard";
+import Filter from "./Filter/Filter";
+import Nav from "./Nav/Nav";
+import ProductCard from "./ProductCard/ProductCard";
+
+export { CategoryCard, CartCard, Filter, Nav, ProductCard };
