+++ src/Pages/index.js
@@ -1,0 +1,6 @@
+import Cart from "./Cart/Cart";
+import Home from "./Home/Home";
+import ProductList from "./ProductList/ProductList";
+import Wishlist from "./Wishlist/Wishlist";
+
+export { Cart, Home, ProductList, Wishlist };
